added rest endpoints

+++ salesapp/src/main/java/com/sparksupport/salesapp/service/ProductService.java
@@ -8,13 +8,13 @@ public interface ProductService {
 
     public List<Product> getAllProducts() ;
 
-    public Product getProductById(int id) ;
+    public Product getProductById(Long id) ;
 
     public void addProduct(Product product) ;
 
-    public void updateProduct(int id, Product updatedProduct) ;
+    public void updateProduct(Long id, Product updatedProduct) ;
 
-    public void deleteProduct(int id) ;
+    public void deleteProduct(Long id) ;
     public double getRevenueByProduct(Long productId);
     public double getTotalRevenue() ;
 }
